[REFACTOR] #269: provier 파일 정리
