+++ frontend/src/context/ModalProviders/QuoteSummaryModalProvider.tsx
