+++ frontend/src/utils/apis.ts
@@ -10,3 +10,4 @@ export const INNER_COLOR_API = `${BASE_API_URL}/api/cars/colors/inner`;
 export const DEFAULT_INFO_API = `${BASE_API_URL}/api/cars/infos/defaults`;
 export const SHARE_INFO_API = `${BASE_API_URL}/api/quote/infos/shares`;
 export const CAR_LIST_API = `${BASE_API_URL}/api/cars/list`;
+export const QUOTE_LIST_API = `${BASE_API_URL}/api/quotes/list`;
